+++ .coveragerc
@@ -1,3 +1,3 @@
 [run]
 omit =
-    *custom*
+    *custo*
