+++ setup.cfg
@@ -1,2 +1,2 @@
 [metadata]
-description-file = README.rst
+description-file = README.rstt
