+++ src/oemof/solph/groupings.py
