+++ src/oemof/solph/models.py
@@ -10,9 +10,8 @@ SPDX-License-Identifier: MIT
 import pyomo.environ as po
 from pyomo.opt import SolverFactory
 from pyomo.core.plugins.transform.relax_integrality import RelaxIntegrality
-from oemof.solph import blocks
+from oemof.solph import blocks, processing
 from oemof.solph.plumbing import sequence
-from oemof.outputlib import processing
 import warnings
 import logging
