+++ tests/test_models.py
@@ -21,7 +21,7 @@ import warnings
 def test_timeincrement_with_valid_timeindex():
     datetimeindex = pd.date_range('1/1/2012', periods=1, freq='H')
     es = solph.EnergySystem(timeindex=datetimeindex)
-    m = solph.models.BaseModel(es)
+    m = oemof.solph.models.BaseModel(es)
     eq_(m.timeincrement[0], 1)
     eq_(es.timeindex.freq.nanos / 3.6e12, 1)
 
@@ -29,18 +29,18 @@ def test_timeincrement_with_valid_timeindex():
 @raises(AttributeError)
 def test_timeincrement_with_non_valid_timeindex():
     es = solph.EnergySystem(timeindex=4)
-    solph.models.BaseModel(es)
+    oemof.solph.models.BaseModel(es)
 
 
 def test_timeincrement_value():
     es = solph.EnergySystem(timeindex=4)
-    m = solph.models.BaseModel(es, timeincrement=3)
+    m = oemof.solph.models.BaseModel(es, timeincrement=3)
     eq_(m.timeincrement[0], 3)
 
 
 def test_timeincrement_list():
     es = solph.EnergySystem(timeindex=4)
-    m = solph.models.BaseModel(es, timeincrement=[0, 1, 2, 3])
+    m = oemof.solph.models.BaseModel(es, timeincrement=[0, 1, 2, 3])
     eq_(m.timeincrement[3], 3)
 
 
@@ -94,7 +94,7 @@ def test_optimal_solution():
     es.add(solph.Sink(inputs={bel: solph.Flow(
         nominal_value=5, actual_value=[1], fixed=True)}))
     es.add(solph.Source(outputs={bel: solph.Flow(variable_costs=5)}))
-    m = solph.models.Model(es, timeincrement=1)
+    m = oemof.solph.models.Model(es, timeincrement=1)
     m.solve('cbc')
     m.results()
     outputlib.processing.meta_results(m)
@@ -110,7 +110,7 @@ def test_infeasible_model():
                 nominal_value=5, actual_value=[1], fixed=True)}))
             es.add(solph.Source(outputs={bel: solph.Flow(
                 nominal_value=4, variable_costs=5)}))
-            m = solph.models.Model(es, timeincrement=1)
+            m = oemof.solph.models.Model(es, timeincrement=1)
             m.solve(solver='cbc')
             assert "Optimization ended with status" in str(w[0].message)
             outputlib.processing.meta_results(m)
