+++ tests/test_scripts/test_solph/test_generic_chp/test_generic_chp.py
@@ -17,7 +17,8 @@ import os
 import pandas as pd
 import oemof.solph as solph
 from oemof.network import Node
-from oemof.outputlib import processing, views
+from oemof.outputlib import views
+from oemof.solph import processing
 
 
 def test_gen_chp():
