+++ tests/test_scripts/test_solph/test_storage_investment/test_invest_storage_regression.py
@@ -11,7 +11,8 @@ SPDX-License-Identifier: MIT
 
 import oemof.solph as solph
 from oemof.network import Node
-from oemof.outputlib import processing, views
+from oemof.outputlib import views
+from oemof.solph import processing
 
 import logging
 import pandas as pd
