+++ tests/test_scripts/test_solph/test_storage_investment/test_storage_investment.py
@@ -42,7 +42,8 @@ from oemof.tools import economics
 
 import oemof.solph as solph
 from oemof.network import Node
-from oemof.outputlib import processing, views
+from oemof.outputlib import views
+from oemof.solph import processing
 
 import logging
 import os
