+++ tests/test_scripts/test_solph/test_storage_investment/test_storage_with_tuple_label.py
@@ -39,7 +39,8 @@ from collections import namedtuple
 
 import oemof.solph as solph
 from oemof.network import Node
-from oemof.outputlib import processing, views
+from oemof.outputlib import views
+from oemof.solph import processing
 
 import logging
 import os
